Previous Rounds and Map changes
The previous rounds page has been added which display the rounds previously played by the user, allowing them to edit them and discard them. The map page has been changed to have a second nav bar (at the top, old one moved to the bottom). This allows the user to finish the round, return to home, see hole information, and select to jump to any hole in the course. This nav bar has also been added to the Score page.